requirements.txt
fastapi
groq
python-dotenv
gtts
uvicorn
python-multipart